requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.17.10
botocore==1.20.10
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.6
django-admin-interface==0.15.3
django-apptemplates==1.5
django-audiofield==0.10.0
django-baton==2.0.0
django-braces==1.14.0
django-celery==3.1.17
django-cleanup==5.1.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-extensions==3.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.14.3
django-heroku==0.3.1
django-material-admin==1.7.16
django-phonenumber-field==5.0.0
django-smart-selects==1.5.9
django-sslserver==0.22
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.5
drf-yasg==1.20.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.17
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==5.0.1
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.4.0
sorl-thumbnail==12.7.0
sorl-thumbnail-serializer-field==0.2.1
sox==1.4.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.1.0
